Fix `newAdaptedPromise<Promise<T>, Adapter>()`.
This was failing to chain the promises, and so returning `Promise<Promise<T>>`.

The idea here is you can create a PromiseAdapter which eventually produces another promise to chain to. The adapter is finished and should be destroyed at that point, but the final promise should then redirect to the new promise.